requirements.txt
pyperclip>=1.6.0
requests>=2.18.4
adal>=1.0.0
PyPDF2
reportlab
Pillow
flask